+++ GitTest/src/main/java/com/smhrd3/model/SNSDTO.java
@@ -2,13 +2,13 @@ package com.smhrd3.model;
 
 public class SNSDTO {
 	 // 소셜 순번 
-    public Double sns_seq;
+    public int sns_seq;
 
     // 기준 연월 
     public String sns_yearmonth;
 
     // 검색량 
-    public Double sns_search;
+    public int sns_search;
 
     // 지역 
     public String sns_are;
@@ -16,11 +16,11 @@ public class SNSDTO {
     // 관리자 아이디 
     public String admin_id;
 
-	public Double getSns_seq() {
+	public int getSns_seq() {
 		return sns_seq;
 	}
 
-	public void setSns_seq(Double sns_seq) {
+	public void setSns_seq(int sns_seq) {
 		this.sns_seq = sns_seq;
 	}
 
@@ -32,11 +32,11 @@ public class SNSDTO {
 		this.sns_yearmonth = sns_yearmonth;
 	}
 
-	public Double getSns_search() {
+	public int getSns_search() {
 		return sns_search;
 	}
 
-	public void setSns_search(Double sns_search) {
+	public void setSns_search(int sns_search) {
 		this.sns_search = sns_search;
 	}
 
@@ -56,7 +56,7 @@ public class SNSDTO {
 		this.admin_id = admin_id;
 	}
 
-	public SNSDTO(Double sns_seq, String sns_yearmonth, Double sns_search, String sns_are, String admin_id) {
+	public SNSDTO(int sns_seq, String sns_yearmonth, int sns_search, String sns_are, String admin_id) {
 		this.sns_seq = sns_seq;
 		this.sns_yearmonth = sns_yearmonth;
 		this.sns_search = sns_search;
